Working on Ring Management
Added a page to create rings, view rings, and manage individual rings. The View Rings page is functional enough, but the circle drawing should probably be done in JavaScript. The Manage Ring page is a WIP and the Create Ring page is blank. Also, the website seems to have grown fond of Natalie Portman.